Cleanup ChallengeProgressActivityStarter usage (#3845)
Move `ChallengeProgressActivityStarter` and implementation out of
`StripePaymentController` to `Stripe3DS2Authenticator`. It is only used
in `Stripe3DS2Authenticator`.
