+++ payments-core/src/main/java/com/stripe/android/payments/core/authentication/Stripe3DS2Authenticator.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.payments.core.authentication
 
+import android.content.Context
 import androidx.annotation.VisibleForTesting
 import com.stripe.android.PaymentAuthConfig
 import com.stripe.android.PaymentRelayStarter
@@ -16,11 +17,14 @@ import com.stripe.android.networking.ApiRequest
 import com.stripe.android.networking.StripeRepository
 import com.stripe.android.payments.DefaultStripeChallengeStatusReceiver
 import com.stripe.android.payments.Stripe3ds2CompletionStarter
+import com.stripe.android.stripe3ds2.init.ui.StripeUiCustomization
 import com.stripe.android.stripe3ds2.service.StripeThreeDs2Service
 import com.stripe.android.stripe3ds2.transaction.ChallengeParameters
 import com.stripe.android.stripe3ds2.transaction.MessageVersionRegistry
+import com.stripe.android.stripe3ds2.transaction.SdkTransactionId
 import com.stripe.android.stripe3ds2.transaction.Stripe3ds2ActivityStarterHost
 import com.stripe.android.stripe3ds2.transaction.Transaction
+import com.stripe.android.stripe3ds2.views.ChallengeProgressActivity
 import com.stripe.android.view.AuthActivityStarterHost
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.withContext
@@ -40,11 +44,11 @@ internal class Stripe3DS2Authenticator(
     private val analyticsRequestFactory: AnalyticsRequestFactory,
     private val threeDs2Service: StripeThreeDs2Service,
     private val messageVersionRegistry: MessageVersionRegistry,
-    private val challengeProgressActivityStarter: StripePaymentController.ChallengeProgressActivityStarter,
     private val stripe3ds2Config: PaymentAuthConfig.Stripe3ds2Config,
     private val stripe3ds2CompletionStarterFactory: (AuthActivityStarterHost, Int) -> Stripe3ds2CompletionStarter,
     private val workContext: CoroutineContext,
-    private val uiContext: CoroutineContext
+    private val uiContext: CoroutineContext,
+    private val challengeProgressActivityStarter: ChallengeProgressActivityStarter = DefaultChallengeProgressActivityStarter()
 ) : IntentAuthenticator {
 
     override suspend fun authenticate(
@@ -354,6 +358,34 @@ internal class Stripe3DS2Authenticator(
                     workContext = workContext
                 ),
                 maxTimeout
+            )
+        }
+    }
+
+    internal fun interface ChallengeProgressActivityStarter {
+        fun start(
+            context: Context,
+            directoryServerName: String,
+            cancelable: Boolean,
+            uiCustomization: StripeUiCustomization,
+            sdkTransactionId: SdkTransactionId
+        )
+    }
+
+    internal class DefaultChallengeProgressActivityStarter : ChallengeProgressActivityStarter {
+        override fun start(
+            context: Context,
+            directoryServerName: String,
+            cancelable: Boolean,
+            uiCustomization: StripeUiCustomization,
+            sdkTransactionId: SdkTransactionId
+        ) {
+            ChallengeProgressActivity.show(
+                context,
+                directoryServerName,
+                cancelable,
+                uiCustomization,
+                sdkTransactionId
             )
         }
     }
